Move pluggbale formula to pkg/marketmaking
